# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da GateWay API GymApp",
    "description": "Documentação da gateway api de suporte à aplicação GymApp",
    "contact": { "name": "Teste", "email": "teste@teste.com" },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:2900", "description": "API de teste" }
  ],
  "paths": {
...
    "/aluno/agenda/avaliacao/": {
      "post": {
        "summary": "Registo de pedidos de agendamentos de avaliação",
        "description": "Esta rota vai registar na base de dados um agendamento de avaliações",
        "tags": ["Agenda"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ginasioId": { "type": "string" },
                  "dataAgendamento": { "type": "string" }
                }
              },
              "examples": {
                "Inserir Agendamento de Avaliacoes": {
                  "value": {
                    "ginasioId": "531316a4-b454-42ec-bc01-7358a8c3bc48",
                    "dataAgendamento": "2022-05-03T13:24:10.574Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Inserir Agendamento": {
                    "value": { "msg": "Agendamento inserido com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/aluno/agenda/desafios": {
      "post": {
        "summary": "Registo de pedidos de agendamentos de desafios",
        "description": "Esta rota vai registar na base de dados um agendamento de desafios",
        "tags": ["Agenda"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "desafioId": { "type": "string" },
                  "ginasioId": { "type": "string" },
                  "dataAgendamento": { "type": "string" }
                }
              },
              "examples": {
                "Inserir Agendamento de Desafios": {
                  "value": {
                    "desafioId": "05e05f72-face-4180-8289-b0707684e667",
                    "ginasioId": "531316a4-b454-42ec-bc01-7358a8c3bc48",
                    "dataAgendamento": "2022-05-03T13:24:10.574Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Inserir Agendamento": {
                    "value": { "msg": "Agendamento inserido com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/avaliacoes/": {
      "get": {
        "summary": "Ver avaliações",
        "description": "Esta rota vai apresentar todas as avaliações",
        "tags": ["Avaliações"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "Obter Avaliações": {
                    "value": [
                      {
                        "avaliacao_id": "640e1f20-938f-4ed6-8ef4-7b02ab98b486",
                        "data": "2022-04-18T15:06:37.318Z",
                        "peso": 55,
                        "musculo": 80,
                        "gordura_corporal": 1,
                        "gordura_visceral": 1,
                        "agua": 30,
                        "proteina": 20,
                        "massa_ossea": 10,
                        "metabolismo_basal": 1600,
                        "avaliacao_imagens": [],
                        "medidas_avaliacao": [],
                        "users_avaliacoes_treinador_idTousers": {
                          "nome": "treinador",
                          "imagem_url": null
                        }
                      },
                      {
                        "avaliacao_id": "89bb2521-b35e-42ac-beb7-bf0485a415d7",
                        "data": "2022-04-18T15:14:52.695Z",
                        "peso": 55,
                        "musculo": 80,
                        "gordura_corporal": 1,
                        "gordura_visceral": 1,
                        "agua": 30,
                        "proteina": 20,
                        "massa_ossea": 10,
                        "metabolismo_basal": 1600,
                        "avaliacao_imagens": [{ "url": "https://imagem.1" }],
                        "medidas_avaliacao": [
                          {
                            "medida": "50",
                            "unidade_medida": "cm",
                            "locais_medidas": {
                              "descricao": "Peito",
                              "unilado": true
                            }
                          },
                          {
                            "medida": "30x30",
                            "unidade_medida": "cm",
                            "locais_medidas": {
                              "descricao": "Braço",
                              "unilado": false
                            }
                          }
                        ],
                        "users_avaliacoes_treinador_idTousers": {
                          "nome": "treinador",
                          "imagem_url": null
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/treinos": {
      "post": {
        "summary": "Criar treino",
        "description": "Esta rota vai criar na base de dados os treinos",
        "tags": ["Treinos"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "treinoId": { "type": "string" },
                  "atividadeId": { "type": "string", "nullable": true },
                  "modalidadeId": { "type": "string", "nullable": true },
                  "duracao": { "type": "string" },
                  "calorias": { "type": "number" },
                  "distancia": { "type": "number" },
                  "data": { "type": "string" }
                }
              },
              "examples": {
                "Criar Treino": {
                  "value": {
                    "treinoId": "48474843-d48b-4786-9678-fe69248923dc",
                    "atividadeId": null,
                    "modalidadeId": "4272f33a-b2c9-46bf-83ab-c8a1a85fbd52",
                    "duracao": "10",
                    "calorias": 100,
                    "distancia": 500,
                    "data": "2022-04-18T17:51:26.567Z",
                    "isDeleted": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Criar treino": {
                    "value": { "message": "Treino criado com sucesso" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Editar treinos",
        "description": "Esta rota vai editar na base de dados um treino de um utilizador",
        "tags": ["Treinos"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "treinoId": { "type": "string" },
                  "atividadeId": { "type": "string", "nullable": true },
                  "modalidadeId": { "type": "string", "nullable": true },
                  "duracao": { "type": "string" },
                  "calorias": { "type": "number" },
                  "distancia": { "type": "number" },
                  "data": { "type": "number" }
                }
              },
              "examples": {
                "Editar Treino": {
                  "value": {
                    "treinoId": "48474843-d48b-4786-9678-fe69248923dc",
                    "atividadeId": "7c98d84f-10d4-44f6-aafa-ddfaa0bee695",
                    "modalidadeId": null,
                    "duracao": "50",
                    "calorias": 20,
                    "distancia": 30,
                    "data": "2022-05-03T17:53:33.437Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Editar Exercicio": {
                    "value": {
                      "treino_id": "bdf6a96c-89ae-4857-9db2-849bf66f6ed7",
                      "uid": "000d1e14-617e-423e-8a1a-f63d4fa5af6a",
                      "atividade_id": "7c98d84f-10d4-44f6-aafa-ddfaa0bee695",
                      "modalidade_id": null,
                      "duracao": "50",
                      "calorias": 20,
                      "distancia": 30,
                      "data": "2022-04-18T17:53:33.437Z",
                      "isDeleted": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/treino/{treinoId}/": {
      "delete": {
        "summary": "Remover treino",
        "description": "Esta rota vai remover da base de dados os treinos criados",
        "tags": ["Treinos"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "treinoId",
            "in": "path",
            "description": "Id do treino",
            "required": true,
            "schema": { "properties": { "treinoId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover treino": {
                    "value": { "msg": "Treino removido com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/agenda/desafios/{agendamentoId}/agendamento/": {
      "delete": {
        "summary": "Remover agendamento de um desafio",
        "description": "Esta rota vai apagar na base de dados um agendamento de um desafio",
        "tags": ["Agenda"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "agendamentoId",
            "in": "path",
            "description": "Id do agendamento",
            "required": true,
            "schema": {
              "properties": { "agendamentoId": { "type": "string" } }
            }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover Agendamento de um Desafio": {
                    "value": {
                      "msg": "Agendamento do desafio removido com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/aluno/agenda/avaliacao/{agendamentoId}/agendamento/": {
      "delete": {
        "summary": "Remover agendamento de uma avaliação",
        "description": "Esta rota vai apagar na base de dados um agendamento de uma avaliação",
        "tags": ["Agenda"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "agendamentoId",
            "in": "path",
            "description": "Id do agendamento",
            "required": true,
            "schema": {
              "properties": { "agendamentoId": { "type": "string" } }
            }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover Agendamento de uma Avaliação": {
                    "value": {
                      "msg": "Agendamento da avaliação removido com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/treinos/": {
      "get": {
        "summary": "Ver todos os treinos de um aluno",
        "description": "Esta rota vai obter todos os treinos de um aluno",
        "tags": ["Treinos"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Todos os Treinos de um Aluno": {
                    "value": [
                      {
                        "treino_id": "98a44040-f84d-44d3-a141-b60c9bbf73c4",
                        "uid": "000d1e14-617e-423e-8a1a-f63d4fa5af6a",
                        "atividade_id": "7c98d84f-10d4-44f6-aafa-ddfaa0bee695",
                        "modalidade_id": null,
                        "duracao": "50",
                        "calorias": 20,
                        "distancia": 30,
                        "data": "2022-04-15T13:56:13.821Z",
                        "isDeleted": false
                      },
                      {
                        "treino_id": "bdf6a96c-89ae-4857-9db2-849bf66f6ed7",
                        "uid": "000d1e14-617e-423e-8a1a-f63d4fa5af6a",
                        "atividade_id": "7c98d84f-10d4-44f6-aafa-ddfaa0bee695",
                        "modalidade_id": null,
                        "duracao": "50",
                        "calorias": 20,
                        "distancia": 30,
                        "data": "2022-04-18T17:53:33.437Z",
                        "isDeleted": false
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/planoTreino/{startDate}/{endDate}": {
      "get": {
        "summary": "Obter plano de treino semanal",
        "description": "Esta rota vai permitir obter plano de treino de uma semana dano o intervalo de tempo",
        "tags": ["Plano Treino"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "description": "data de inicio",
            "required": true,
            "schema": { "properties": { "startDate": { "type": "string" } } },
            "examples": { "startDate": { "value": "2022-04-11" } }
          },
          {
            "name": "endDate",
            "in": "path",
            "description": "data de fim",
            "required": true,
            "schema": { "properties": { "endDate": { "type": "string" } } },
            "examples": { "endDate": { "value": "2022-04-17" } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Obter Treino": {
                    "value": [
                      {
                        "data": "2022-04-14T00:00:00.000Z",
                        "isRealizado": false,
                        "treinador": {
                          "nome": "treinador teste",
                          "email": "treinador@treinador.com",
                          "imagem_url": null
                        },
                        "modalidade": { "nome": "Crossfit" },
                        "bloco_treino": [
                          {
                            "nome": "Aquecimento",
                            "descricao": "5 Rondas",
                            "exercicios_bloco": [
                              {
                                "n_ordem_exercicio": 0,
                                "series_exercicio": [
                                  { "valor": "12", "n_ordem_serie": 0 },
                                  { "valor": "12", "n_ordem_serie": 1 },
                                  { "valor": "12", "n_ordem_serie": 2 },
                                  { "valor": "12", "n_ordem_serie": 3 }
                                ],
                                "exercicio": {
                                  "nome": "Teste1",
                                  "descricao": "Teste2",
                                  "imagens": [
                                    { "url": "url1" },
                                    { "url": "url2" },
                                    { "url": "url3" },
                                    { "url": "url4" }
                                  ],
                                  "musculos": [
                                    {
                                      "musculos": {
                                        "nome": "Bicep",
                                        "img_url": "http://imagem/bicep"
                                      }
                                    },
                                    {
                                      "musculos": {
                                        "nome": "Tricep",
                                        "img_url": "http://imagem/tricep"
                                      }
                                    },
                                    {
                                      "musculos": {
                                        "nome": "Costas",
                                        "img_url": "http://imagem/costas"
                                      }
                                    }
                                  ]
                                }
                              }
                            ]
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/agenda/avaliacoes/": {
      "get": {
        "summary": "Ver todos os pedidos de agendamento de avaliação de um aluno",
        "description": "Esta rota vai obter todos os pedidos de agendamento de avaliação de um aluno",
        "tags": ["Agenda"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Agendamentos de Avaliações de um Aluno": {
                    "value": {
                      "msg": "Ver agendamento de avaliações de um aluno obtido com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/aluno/agenda/desafios/": {
      "get": {
        "summary": "Ver todos os pedidos de agendamento de desafios de um aluno",
        "description": "Esta rota vai obter todos os pedidos de agendamento de desafios de um aluno",
        "tags": ["Agenda"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Ver Agendamentos de Desafios de um Aluno": {
                    "value": {
                      "msg": "Ver agendamento de desafios de um aluno obtido com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/aluno/plano/realizado/": {
      "put": {
        "summary": "Colocar um plano de treino como realizado",
        "description": "Esta rota vai editar na base de dados o estado da realização de um plano de treino",
        "tags": ["Plano Treino"],
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "planoId": { "type": "string" }
                }
              },
              "examples": {
                "Plano Realizado": {
                  "value": {
                    "planoId": "89db473b-69d2-477a-9a83-d77193ea6b86"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Editar Estado da Realização do Plano de Treino": {
                    "value": {
                      "plano_treino_id": "89db473b-69d2-477a-9a83-d77193ea6b86",
                      "aluno_id": "000d1e14-617e-423e-8a1a-f63d4fa5af6a",
                      "treinador_id": "6d112678-f40c-43bc-a80a-866f12fa26f2",
                      "data": "2022-04-18T00:00:00.000Z",
                      "isRealizado": true,
                      "modalidade_id": "4272f33a-b2c9-46bf-83ab-c8a1a85fbd52",
                      "isDeleted": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/aluno/plano/{planoId}/realizado/": {
      "delete": {
        "summary": "Remover o estado realizado de um plano de treino",
        "description": "Esta rota vai remover na base de dados o estado da realização de um plano de treino",
        "tags": ["Plano Treino"],
        "security": [{ "basicAuth": [] }],
        "parameters": [
          {
            "name": "planoId",
            "in": "path",
            "description": "Id do plano de treino",
            "required": true,
            "schema": { "properties": { "planoId": { "type": "string" } } }
          }
        ],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "properties": { "msg": { "type": "string" } } },
                "examples": {
                  "Remover Estado da Realização do Plano de Treino": {
                    "value": {
                      "msg": "Remover estado da realização do plano de treino aceite com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "basicAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "uid": { "type": "string" },
          "email": { "type": "string" },
          "nome": { "type": "string" },
          "password": { "type": "string" },
          "data_nasc": { "type": "string" },
          "hashtag": { "type": "string" },
          "data_entrada": { "type": "string" },
          "funcao_id": { "type": "string" },
          "refresh_token": { "type": "string" },
          "genero": { "type": "number" },
          "pontos": { "type": "number" },
          "descricao": { "type": "string" },
          "imagem_url": { "type": "string" },
          "isDeleted": { "type": "boolean" }
        }
      },
...
    }
  }
}